Fill list of completion data on demand
Instead of initalizing all the data as soon as anything is requested,
only initialize the lists that are actually needed. Do this by
introducing a "getter" for the lists instead of referencing s:cmake_foo
directly. The getter checks if the list is empty and then fills it. The
many more calls to Version() is mitigated by the previous commit.

Currently this achieves nothing of value for the Help() function, but
does no harm either.